beta eval example
To make this I had to factor the example expr code to allow for the more
complex expression used in the example. I also had to add something to
the EXPR_APPLY_LAMBDA concat rule so that it would store the substituted
function arguments in the right place, that I forgot to do the first
time since I had no easy way to test the code.

This will provide a nice (though cumbersome) way of testing beta
reduction, as I implement it, since beta evaluation allows this
expression to type check.