heterogeneous equality
These builtins are rather magical in that their types are specified in
terms of themselves... But that is just to save us from having two sets
of builtins, one for homogeneous and one for heterogeneous equality,
thus eliminating the need for more builtins to convert between those
builtins. Either way of doing things seems fine, though.