fix arg count calculation
It turns out it wasn't a bug that I fixed in the last commit, just a
weird convention that I didn't understand. I should probably change it
so that concat_intros_... does change the arg count, since that is
apparently what seemed more logical when I came back to this code today.